New structure created for multiple lambdas creation
